move script to the correct location
